Agregada funcionalidad de Deposito en Movimientos
- Modificada la funcionalidad de retiro para que verifique que tenga
saldo suficiente para retirar

+++ BancApp/src/bancapp/daos/interfaces/IChequeraDAO.java
@@ -14,7 +14,7 @@ public interface IChequeraDAO {
   
   public String insertarChequera(Chequera chequera) throws Exception;
   
-  public Chequera consultarChequera(int idChequera) throws Exception;
+  public Chequera consultarChequera(long idChequera) throws Exception;
   
   public String modificarChequera(Chequera chequera) throws Exception;
