+++ BancApp/src/bancapp/services/interfaces/IChequeraService.java
@@ -15,7 +15,7 @@ public interface IChequeraService {
   
   public String insertarChequera(Chequera chequera) throws Exception;
   
-  public Chequera consultarChequera(int idChequera) throws Exception;
+  public Chequera consultarChequera(long idChequera) throws Exception;
   
 //  public String modificarChequera(Chequera chequera) throws Exception;
